requirements.txt
python-telegram-bot==13.11
python-dotenv==0.21.0
requests==2.31.0